Se agregan las funciones faltantes para crud, Crear, Leer, Actuliazar y Eliminar, tabla con las visas y botones en correcto funcionamiento

+++ SGPI/Models/Homologacion.cs
@@ -11,7 +11,7 @@ public partial class Homologacion
 
     public int Idasig { get; set; }
 
-    public int Iduser { get; set; }
+    public long Iduser { get; set; }
 
     public int Tipodoc { get; set; }
